should move cache into note editor for better code quality. not regist all the things in bese class.
