+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Rows/Extends/Notes/NoteEditor.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using System.Collections.Generic;
 using System.Linq;
 using osu.Framework.Allocation;
 using osu.Framework.Bindables;
@@ -10,6 +11,9 @@ using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Timing;
 using osu.Game.Rulesets.Karaoke.UI.Position;
 using osu.Game.Rulesets.Karaoke.UI.Scrolling;
+using osu.Game.Rulesets.Timing;
+using osu.Game.Rulesets.UI.Scrolling;
+using osu.Game.Rulesets.UI.Scrolling.Algorithms;
 using osu.Game.Screens.Edit;
 
 namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Extends.Notes
@@ -21,6 +25,9 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Extends.Notes
 
         [Cached(Type = typeof(INotePositionInfo))]
         private readonly PreviewNotePositionInfo notePositionInfo = new PreviewNotePositionInfo();
+
+        [Cached(Type = typeof(IScrollingInfo))]
+        private readonly LocalScrollingInfo scrollingInfo = new LocalScrollingInfo();
 
         private readonly Lyric lyric;
 
@@ -38,14 +45,15 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Extends.Notes
                 Children = new Drawable[]
                 {
                     // layers below playfield
-                    Playfield = new EditorNotePlayfield(columns),
+                    Playfield = new EditorNotePlayfield(columns)
+                    {
+                        // set stop clock and navigation to target time.
+                        Clock = new StopClock(lyric.LyricStartTime)
+                    },
                     // layers above playfield
                     new EditNoteBlueprintContainer(lyric),
                 }
             };
-
-            // set stop clock and navigation to target time.
-            Playfield.Clock = new StopClock(lyric.LyricStartTime);
         }
 
         [BackgroundDependencyLoader]
@@ -65,6 +73,19 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Extends.Notes
             public IBindable<NotePositionCalculator> Position { get; } = new Bindable<NotePositionCalculator>(new NotePositionCalculator(columns, 12, ScrollingNotePlayfield.COLUMN_SPACING));
 
             public NotePositionCalculator Calculator => Position.Value;
+        }
+
+        private class LocalScrollingInfo : IScrollingInfo
+        {
+            public IBindable<ScrollingDirection> Direction { get; } = new Bindable<ScrollingDirection>(ScrollingDirection.Left);
+
+            public IBindable<double> TimeRange { get; } = new BindableDouble(5000)
+            {
+                MinValue = 1000,
+                MaxValue = 10000
+            };
+
+            public IScrollAlgorithm Algorithm { get; } = new SequentialScrollAlgorithm(new List<MultiplierControlPoint>());
         }
     }
 }
